correcion de errores para traer los dactos reales en SelectNewProducActivity

+++ app/src/main/java/com/example/comfyapp/Product.kt
@@ -5,7 +5,7 @@ data class Product(
     val name: String,
     val price: Double,
     val imageBase64: String? = null,
-    val stock: Int = 0,
+    val stock: Double = 0.0,
     val description: String? = null,
     var lastEntryQty: Double = 0.0,
     var lastEntryDate: String = ""
